services/api/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic[email]==2.5.0
sqlalchemy==2.0.23
alembic==1.13.1
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
email-validator==2.1.0
python-dotenv==1.0.0
redis==5.0.1
celery==5.3.4
firebase-admin==6.3.0
svix==1.20.0
httpx==0.25.2
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.12.0
flake8==7.0.0
mypy==1.7.1
requests==2.31.0
jinja2==3.1.2
aiofiles==23.2.1
fastapi-users==12.1.3
fastapi-users[sqlalchemy]==12.1.3